Backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
googlemaps==4.10.0
geopy==2.4.1
google-generativeai==0.3.2
werkzeug==3.0.1
PyJWT==2.8.0
